pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.12",
    "click>=8.2.1",
    "markdown>=3.8",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "pdfplumber>=0.11.6",
    "pydantic>=2.11.5",
    "pypdf2>=3.0.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
